Changes Points to double?
Will still display as int, but this allows point sharing if the committee agree with this approach.

PS C:\repos\wvcc\WellandValleyCC.github.io\processor\ClubProcessor> dotnet ef migrations add UpdatePointsToDouble --context EventDbContext
Build started...
Build succeeded.
An operation was scaffolded that may result in the loss of data. Please review the migration for accuracy.
Done. To undo this action, use 'ef migrations remove'
PS C:\repos\wvcc\WellandValleyCC.github.io\processor\ClubProcessor> dotnet ef database update --context EventDbContext
Build started...
Build succeeded.
Acquiring an exclusive lock for migration application. See https://aka.ms/efcore-docs-migrations-lock for more information if this takes too long.
Applying migration '20251031092012_UpdatePointsToDouble'.
The migration operation 'PRAGMA foreign_keys = 0;
' from migration 'UpdatePointsToDouble' cannot be executed in a transaction. If the app is terminated or an unrecoverable error occurs while this operation is being executed then the migration will be left in a partially applied state and would need to be reverted manually before it can be applied again. Create a separate migration that contains just this operation.
Done.

+++ processor/ClubProcessor/Models/Ride.cs
@@ -41,10 +41,10 @@ namespace ClubProcessor.Models
             get => EventPosition;
             set => EventPosition = value;
         }
-        public int? SeniorsPoints { get; set; }
+        public double? SeniorsPoints { get; set; }
 
         public int? WomenPosition { get; set; }
-        public int? WomenPoints { get; set; }
+        public double? WomenPoints { get; set; }
 
         [NotMapped]
         public TimeSpan? VeteransStandardTime { get; set; }
@@ -53,37 +53,37 @@ namespace ClubProcessor.Models
         public TimeSpan? VeteransTimeDelta { get; set; }
 
         public int? VeteransPosition { get; set; }
-        public int? VeteransPoints { get; set; }
+        public double? VeteransPoints { get; set; }
 
         public int? RoadBikeMenPosition { get; set; }
-        public int? RoadBikeMenPoints { get; set; }
+        public double? RoadBikeMenPoints { get; set; }
 
         public int? RoadBikeWomenPosition { get; set; }
-        public int? RoadBikeWomenPoints { get; set; }
+        public double? RoadBikeWomenPoints { get; set; }
 
         public int? JuvenilesPosition { get; set; }
-        public int? JuvenilesPoints { get; set; }
+        public double? JuvenilesPoints { get; set; }
 
         public int? JuniorsPosition { get; set; }
-        public int? JuniorsPoints { get; set; }
+        public double? JuniorsPoints { get; set; }
 
         public int? PremPosition { get; set; }
-        public int? PremPoints { get; set; }
+        public double? PremPoints { get; set; }
 
         public int? League1Position { get; set; }
-        public int? League1Points { get; set; }
+        public double? League1Points { get; set; }
 
         public int? League2Position { get; set; }
-        public int? League2Points { get; set; }
+        public double? League2Points { get; set; }
 
         public int? League3Position { get; set; }
-        public int? League3Points { get; set; }
+        public double? League3Points { get; set; }
 
         public int? League4Position { get; set; }
-        public int? League4Points { get; set; }
+        public double? League4Points { get; set; }
 
         public int? NevBrooksPosition { get; set; }
-        public int? NevBrooksPoints { get; set; }
+        public double? NevBrooksPoints { get; set; }
 
         public int? NevBrooksSecondsGenerated { get; set; }     // Raw handicap value
         public int? NevBrooksSecondsApplied { get; set; }       // Capped or adjusted handicap
